+++ works/scnreader_view.cpp
@@ -44,7 +44,7 @@ void scnreader_view::paintGL()
     for(int j=0;j<this->scnreaderFond.getClouds().size();j++)
         for (int i=0;i<this->scnreaderFond.getClouds(j)->points.size();i++){
 
-            glColor3f(1.0/(float)(j+1),1.0/(float)(this->scnreaderFond.getClouds().size()-j),1.0);
+            glColor3f(1.0/(double)(j+1),1.0/(double)(this->scnreaderFond.getClouds().size()-j),1.0);
             //normalizes points with model->max of cordinates previously finded
             glVertex3f(this->scnreaderFond.getClouds(j)->points[i].x/scnreaderFond.getMaxX(),
                        this->scnreaderFond.getClouds(j)->points[i].y/scnreaderFond.getMaxY(),
